fixes for left-side menu items
